requirements.txt
# requirements.txt
asgiref==3.5.2
bleach==5.0.1
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-ipware==4.0.2
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-api-logger==1.1.10
mysqlclient==2.1.1
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1